Update HBI (#12)
Moved code to 10ay from 10ax.
Added code to update the HBI API calls to the 254 API.
Added Constants file, and moved all of the connection related values to
a single file to make changing servers easier.
Pretty sure Pulse is not working because of a missing user or subscriber
id related to the change to 10ay.

See change in Header.tsx, I think there was a bug here. I changed it to
match my perception of how this is supposed to work.

Hopefully the 30 days back filters work as you expect them to.

---------

Co-authored-by: Shern Jauhal <sjauhal@salesforce.com>